scripts/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pillow==10.4.0
pydantic==2.8.2
